Parallel Whale Optimization Algorithm DV-Hop
This is a Matlab code about applying Parallel WOA to enhance the localization accuracy of DV-Hop of WSN.  The related paper is contained in the zip file. In this paper, two novel communication strategies are proposed and improves the searchability of the original WOA. The parallel can significantly improve the performance of the heuristic algorithm.